Added detailed guide on the internship process to resources
This briefly touches on all of the following: various types of interviews,
practice methods, reaching out to recruiters, more detailed resume advice,
and other resources. I wrote this post to help some people out and have
been giving it to people who've asked me, but I think it could be really
useful for more NYU students to have so I'm making a PR to add this to the
wiki.